Created Cameras subpackage.
