+++ napari_live_recording/__init__.py
@@ -1,6 +1,3 @@
-try:
-    from ._version import version as __version__
-except ImportError:
-    __version__ = "unknown"
+__version__ = "0.1.0"
 
 from ._dock_widget import napari_experimental_provide_dock_widget
