+++ napari_live_recording/_dock_widget.py
@@ -5,215 +5,7 @@ from napari_plugin_engine import napari_hook_implementation
 import numpy as np
 from qtpy.QtWidgets import QWidget, QGridLayout, QPushButton
 from imageio import mimwrite
-
-# Camera classes import
-from abc import ABC, abstractmethod
-import cv2
-from platform import system
-from time import sleep
-
-# Ximea camera support only provided by downloading the Ximea Software package
-# see https://www.ximea.com/support/wiki/apis/APIs for more informations
-from ximea.xiapi import Camera as XiCamera, Xi_error
-from ximea.xiapi import Image as XiImage
-
-CAM_TEST   = "Widget dummy camera"
-CAM_OPENCV = "Default Camera (OpenCV)"
-CAM_XIMEA  = "Ximea xiB-64"
-
-class CameraError(Exception):
-    def __init__(self, error: str) -> None:
-        self.error_description = error
-    
-    def __str__(self) -> str:
-        return self.error_description
-
-class ICamera(ABC):
-    @abstractmethod
-    def __init__(self) -> None:
-        super().__init__()
-        self.camera_name = "ICamera"
-        self.roi = []
-
-    @abstractmethod
-    def __del__(self) -> None:
-        return super().__del__()
-
-    @abstractmethod
-    def __del__(self) -> None:
-        pass
-
-    @abstractmethod
-    def open_device(self) -> bool:
-        pass
-    
-    @abstractmethod
-    def close_device(self) -> None:
-        pass
-
-    @abstractmethod
-    def capture_image(self) -> np.array:
-        pass
-
-    @abstractmethod
-    def set_exposure(self, exposure) -> None:
-        pass
-
-    @abstractmethod
-    def set_roi(self, roi : list) -> None:
-        self.roi = roi
-
-    @abstractmethod
-    def get_roi(self) -> list:
-        return self.roi
-
-    def get_name(self) -> str:
-        return self.camera_name
-
-class TestCamera(ICamera):
-    def __init__(self) -> None:
-        super().__init__()
-        self.camera_name = CAM_TEST
-        self.roi = [500, 500]
-    
-    def __del__(self) -> None:
-        return super().__del__()
-
-    def open_device(self) -> bool:
-        print("Dummy camera opened!")
-        return True
-    
-    def close_device(self) -> None:
-        print("Dummy camera closed!")
-    
-    def capture_image(self) -> np.array:
-        print("Acquiring dummy image!")
-        return np.random.randint(low=0, high=2**8, size=tuple(self.roi), dtype="uint8")
-    
-    def set_exposure(self, exposure) -> None:
-        print(f"Dummy camera exposure set to {exposure}")
-
-    def set_roi(self, roi : list) -> None:
-        self.roi = roi
-    
-    def get_roi(self) -> list:
-        return self.roi
-
-class CameraOpenCV(ICamera):
-    def __init__(self) -> None:
-        super().__init__()
-        self.camera_idx = 0
-        self.camera_api = cv2.CAP_ANY
-        self.camera = None
-        self.camera_name = CAM_OPENCV
-
-        # Windows platforms support discrete exposure times
-        # These are mapped using a dictionary
-        self.exposure_dict = {
-            "1 s"      :  0,
-            "500 ms"   : -1,
-            "250 ms"   : -2,
-            "125 ms"   : -3,
-            "62.5 ms"  : -4,
-            "31.3 ms"  : -5,
-            "15.6 ms"  : -6,
-            "7.8 ms"   : -7,
-            "3.9 ms"   : -8,
-            "2 ms"     : -9,
-            "976.6 us" : -10,
-            "488.3 us" : -11,
-            "244.1 us" : -12,
-            "122.1 us" : -13
-        }
-
-    def __del__(self) -> None:
-        if self.camera is not None:
-            self.camera.release()
-    
-    def __str__(self) -> str:
-        return CAM_OPENCV
-    
-    def open_device(self) -> bool:
-        if self.camera is None:
-            self.camera = cv2.VideoCapture(self.camera_idx, self.camera_api)
-            return self.camera.isOpened()
-        return self.camera.open(self.camera_idx, self.camera_api)
-    
-    def close_device(self) -> None:
-        self.camera.release()
-    
-    def capture_image(self) -> np.array:
-        _ , img = self.camera.read()
-        img = cv2.cvtColor(img, cv2.COLOR_BGR2GRAY)
-        cv2.waitKey(1)
-        return img
-
-    def set_exposure(self, exposure) -> None:
-        if system() == "Windows":
-            exposure = self.exposure_dict[exposure]
-        self.camera.set(cv2.CAP_PROP_EXPOSURE, exposure)
-
-class CameraXimea(ICamera):
-    def __init__(self) -> None:
-        super().__init__()
-        self.camera = XiCamera()
-        self.image = XiImage()
-        self.camera_name = CAM_XIMEA
-        self.roi = [500, 500]
-    
-    def __del__(self) -> None:
-        self.close_device()
-    
-    def __str__(self) -> str:
-        return CAM_XIMEA
-    
-    def open_device(self) -> bool:
-        try:
-            self.camera.open_device()
-            max_lut_idx = self.camera.get_LUTIndex_maximum()
-            for idx in range(0, max_lut_idx):
-                self.camera.set_LUTIndex(idx)
-                self.camera.set_LUTValue(idx)
-            self.camera.enable_LUTEnable()
-            self.camera.start_acquisition()
-        except Xi_error:
-            return False
-        return True
-
-    def close_device(self) -> None:
-        try:
-            self.camera.stop_acquisition()
-            self.camera.close_device()
-            del self.camera
-        except Xi_error: # Camera not connected or already closed
-            pass
-    
-    def capture_image(self) -> np.array:
-        try:
-            self.camera.get_image(self.image)
-            data = self.image.get_image_data_numpy()
-        except Xi_error:
-            data = None
-        sleep(0.01)
-        return data
-
-    def set_exposure(self, exposure) -> None:
-        try:
-            self.camera.set_exposure(exposure)
-        except Xi_error:
-            pass
-    
-    def set_roi(self, roi : list) -> None:
-        self.roi = roi
-
-    def get_roi(self) -> list:
-        return self.roi
-
-supported_cameras = {
-    CAM_TEST   : TestCamera,
-    CAM_OPENCV : CameraOpenCV,
-    CAM_XIMEA  : CameraXimea
-}
+from napari_live_recording.Cameras import *
 
 def acquire(camera : ICamera):
     """
